[refactor] modal 스타일 파일 경로 수정 및 max-height  설정 close #95
